services/notification-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
kafka-python==2.0.2
httpx==0.27.0
